# apidocs.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Kamar API",
      "description": "API untuk mengelola informasi kamar hotel",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      }
    ],
    "tags": [
      {
        "name": "Kamar",
        "description": "API untuk kamar hotel dalam sistem"
      }
    ],
    "paths": {
      "/kamar": {
        "get": {
          "tags": ["Kamar"],
          "summary": "Get all kamar",
          "description": "Get all kamar in the system",
          "operationId": "getKamar",
          "responses": {
            "200": {
              "description": "A list of kamar.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Kamar"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Kamar"],
          "summary": "Add a new kamar",
          "description": "Add a new kamar to the system",
          "operationId": "addKamar",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Kamar"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Kamar added"
            }
          }
        }
      },
      "/kamar/{id}": {
        "get": {
          "tags": ["Kamar"],
          "summary": "Get kamar by id",
          "description": "Get a kamar by its id",
          "operationId": "getKamarById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of kamar to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "A kamar.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Kamar"
                  }
                }
              }
            },
            "404": {
              "description": "Kamar not found"
            }
          }
        },
        "put": {
          "tags": ["Kamar"],
          "summary": "Update an existing kamar",
          "description": "Update an existing kamar in the system",
          "operationId": "updateKamar",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of kamar to update",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Kamar"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Kamar updated"
            },
            "404": {
              "description": "Kamar not found"
            }
          }
        },
        "delete": {
          "tags": ["Kamar"],
          "summary": "Delete a kamar",
          "description": "Delete a kamar in the system",
          "operationId": "deleteKamar",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of kamar to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Kamar deleted"
            },
            "404": {
              "description": "Kamar not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Kamar": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "tipe_kamar": {
              "type": "string"
            },
            "luas": {
              "type": "integer",
              "format": "int32"
            },
            "harga": {
              "type": "integer",
              "format": "int32"
            },
            "isAvailable": {
              "type": "boolean"
            }
          }
        }
      }
    }
  }
  